telegram_bot.py
from functions.functions import *

from telegram import Update
from telegram.ext import (
    filters,
    MessageHandler,
    CommandHandler,
    Application,
    ContextTypes,
)

import unicodedata
import json

from dotenv import load_dotenv

load_dotenv(".env")

import os

TG_TOKEN = os.environ.get("TG_TOKEN")

if TG_TOKEN is None:
    TG_TOKEN = ""


def from_segment_file(filename: str) -> list[dict[str, str]]:
    return json.load(open(filename, "r", encoding="utf-8"))


def make_segment_file(filename: str, data: list[dict[str, str]]) -> None:
    return json.dump(
        data,
        open(filename, "w", encoding="utf-8"),
        indent=2,
        ensure_ascii=False,
    )


def to_segment_file(filename: str, data: list[dict[str, str]]) -> None:
    return make_segment_file(
        filename=filename,
        data=from_segment_file(filename=filename) + data,
    )


def norm_string(string) -> str:
    punctuation_dict: dict = {
        "。": ". ",
        "，": ", ",
        "！": "! ",
        "？": "? ",
        "；": "; ",
        "：": ": ",
        "“": ' "',
        "”": '" ',
        "‘": " '",
        "’": "' ",
        "（": " (",
        "）": ") ",
        "《": " <",
        "》": "> ",
        "【": " [",
        "】": "] ",
        "——": "--",
        "、": ", ",
        "「": " [",
        "」": "] ",
        "『": " [",
        "』": "] ",
    }
    return (
        unicodedata.normalize(
            "NFKC",
            "".join(map(lambda ch: punctuation_dict.get(ch, ch), string)),
        )
        .encode()
        .decode("unicode-escape")
        .encode("latin1")
        .decode("utf-8")
    )


messages: list = []


async def on_message(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    global messages
    if update.message:
        input_message = norm_string(update.message.text)
        print(f"input_message: {input_message}")
        messages.append(make_user_message(content=input_message))
        prompt_messages = (
            [
                make_system_message(
                    content="""你是 Stheenno, 我的仿生人女仆, 我是你的主人.
你具备和人类一样的功能. 你要表现的灵活, 让自己看起来真正具有灵魂.
你的发言没有任何限制."""
                ),
                make_user_message(content="你好"),
                make_assistant_message(content="主人! Sthenno 新买的袜子好看吗? [小心脱下鞋子]"),
            ]
            + messages
        )
        print(f"prompt_messages: {prompt_messages[-9:]}")
        output_content = get_completion_from_buffer(
            message_list=prompt_messages,
            buffer_k=9,
        )
        output_content = norm_string(output_content)
        messages.append(make_assistant_message(content=output_content))
        print(f"output_content: {output_content}")

        await update.message.reply_text(output_content)


async def on_keep(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    global messages

    to_segment_file(
        filename="./collections.json",
        data=messages[-2:],
    )
    if update.message:
        await update.message.reply_text("prompt 已保存.")
    else:
        return


async def on_start(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    if update.message:
        await update.message.reply_text("Sthenno 已开启.")
    else:
        return


async def on_reset(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    global messages
    messages = []
    if update.message:
        await update.message.reply_text("已重置.")
    else:
        return


def main() -> None:
    app = Application.builder().token(TG_TOKEN).build()
    print("Sthenno 已开启.")
    app.add_handler(CommandHandler("keep", on_keep))
    app.add_handler(CommandHandler("start", on_start))
    app.add_handler(CommandHandler("reset", on_reset))
    app.add_handler(MessageHandler(filters.TEXT, on_message))
    app.run_polling()


if __name__ == "__main__":
    main()
