requirements.txt
requests
types-requests
openai
tiktoken
python-dotenv
python-telegram-bot
